requirements.txt
findspark
pyspark==3.2.3
pandas
PyArrow
requests
apache-airflow
